se agregan metodo reporte de cliente

+++ src/main/java/com/citasmedicas/citas/entities/CountClient.java
@@ -4,19 +4,19 @@ import java.util.List;
 
 public class CountClient {
 
-    private int total;
+    private Long total;
     private Client client;
 
-    public CountClient(int total, Client client) {
+    public CountClient(Long total, Client client) {
         this.total = total;
         this.client = client;
     }
 
-    public int getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
